move 최종 폴더 구조
