+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from "react-router-dom";
-import "./App.css";
+
 import router from "./shared/router";
 
 function App() {
