--- src/components/UserComponent/UserComponent.jsx
@@ -1,4 +1,0 @@
-function UserComponent() {
-  return <div>마이페이지</div>;
-}
-export default UserComponent;
