--- src/components/UserComponent/index.js
@@ -1,1 +1,0 @@
-export { default } from "./UserComponent";
