+++ src/pages/MyPage.jsx
@@ -1,0 +1,4 @@
+function MyPage() {
+  return <div>마이페이지</div>;
+}
+export default MyPage;
